# backend/documentation/swagger.yaml
...
swagger: "2.0"
info:
  title: "TRENDMART API"
  description: "API documentation for the TRENDMART application."
  version: "1.0.0"
host: "localhost:8000"
basePath: "/"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# =========================
# Paths
# =========================
paths:
...
  /products/{product_id}/reviews:
    post:
      tags:
        - Reviews
      summary: Add a review for a product
      description: Submit a review for a specific product.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: product_id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/ReviewCreatePayload"
      responses:
        201:
          description: Review created successfully
          schema:
            $ref: "#/definitions/ReviewCreateResponse"
        400:
          description: Validation error
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Product not found
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags:
        - Reviews
      summary: Get all reviews for a product
      description: Retrieve all reviews for a specific product.
      parameters:
        - in: path
          name: product_id
          required: true
          type: integer
      responses:
        200:
          description: Reviews retrieved successfully
          schema:
            type: array
            items:
              $ref: "#/definitions/ReviewResponse"
        404:
          description: Product not found
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      tags:
        - Reviews
      summary: Update a review for a product
      description: Update an existing review for a specific product.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: product_id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/ReviewUpdatePayload"
      responses:
        200:
          description: Review updated successfully
          schema:
            $ref: "#/definitions/ReviewCreateResponse"
        400:
          description: Validation error
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Product or Review not found
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /products/{product_id}/reviews/{review_id}:
    patch:
      tags:
        - Reviews
      summary: Update a specific review for a product
      description: Update an existing review identified by review_id for a specific product.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: product_id
          required: true
          type: integer
        - in: path
          name: review_id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/ReviewUpdatePayload"
      responses:
        200:
          description: Review updated successfully
          schema:
            $ref: "#/definitions/ReviewCreateResponse"
        400:
          description: Validation error
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Product or Review not found
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Reviews
      summary: Delete a review for a product
      description: Delete an existing review for a specific product.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: product_id
          required: true
          type: integer
        - in: path
          name: review_id
          required: true
          type: integer
      responses:
        204:
          description: Review deleted successfully
        404:
          description: Product or Review not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /products/{product_id}/rating_summary:
    get:
      tags:
        - Reviews
      summary: Get rating summary for a product
      description: Retrieve the rating summary (average rating, total reviews) for a specific product.
      parameters:
        - in: path
          name: product_id
          required: true
          type: integer
      responses:
        200:
          description: Rating summary retrieved successfully
          schema:
            $ref: "#/definitions/RatingSummaryResponse"
        404:
          description: Product not found
          schema:
            $ref: "#/definitions/ErrorResponse"

# =========================
# Security Definitions
# =========================
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter 'Bearer' followed by your JWT token. Example: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."

# =========================
# Definitions (Schemas)
# =========================
definitions:
...
  ProductCard:
    type: object
    properties:
      id:
        type: integer
      sku:
        type: string
      name:
        type: string
      description:
        type: string
      price:
        type: number
        format: float
      rating:
        type: number
        format: float
      tags:
        type: array
        items:
          type: string
      primary_image:
        type: string
      thumbnail:
        type: string
      images:
        type: array
        items:
          type: string
      subcategory:
        type: string
      main_category:
        type: string
      category_info:
        type: object
      times_click_on:
        type: integer
      score:
        type: number
        format: float
...
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: "Error message"
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Detailed error message"
      details:
        type: object
        description: "Validation error details"

  # -----------------
  # Payments Schemas
  # -----------------
...
  ReviewCreatePayload:
    type: object
    properties:
      rating:
        type: integer
        minimum: 1
        maximum: 5
      title:
        type: string
      comment:
        type: string
  ReviewUpdatePayload:
    type: object
    properties:
      rating:
        type: integer
        minimum: 1
        maximum: 5
      title:
        type: string
      comment:
        type: string
  ReviewResponse:
    type: object
    properties:
      id:
        type: integer
        example: 12
      cart_id:
        type: integer
        example: 4
      product_id:
        type: integer
        example: 123
      quantity:
        type: integer
        example: 2
      price_per_unit:
        type: number
        format: float
        example: 19.99
...
  ReviewCreateResponse:
    type: object
    properties:
      message:
        type: string
      review:
        $ref: "#/definitions/ReviewResponse"
  RatingSummaryResponse:
    type: object
    properties:
      average_rating:
        type: number
        format: float
      count:
        type: integer
      product:
        $ref: "#/definitions/ProductCard"